+++ src/my_first_adventure_game/game/scenes/__init__.py
@@ -1,11 +1,13 @@
 from my_first_adventure_game.game.scenes.defeat_scene import DefeatScene
 from my_first_adventure_game.game.scenes.gameplay_scene import GameplayScene
+from my_first_adventure_game.game.scenes.pause_scene import PauseScene
 from my_first_adventure_game.game.scenes.title_scene import TitleScene
 from my_first_adventure_game.game.scenes.victory_scene import VictoryScene
 
 __all__ = [
     "DefeatScene",
     "GameplayScene",
+    "PauseScene",
     "TitleScene",
     "VictoryScene",
 ]